Use the `node.name` attribute
Stop introspecting for node names using the private `._obj`
and instead use the public `.name` attr. This returns a file
system path for module items which has to be replaced with a
corresponding `.` delimited module path. With this fix we can
finally collect the `pytest` core test suite :)

Fixes #14
Fixes #6